Bug Tracker Ground Work
Here i have layed down the UI so that i can have a frame of reference to work with, i have worked a couple days on this before making this first commit. I have layed down the ground work for some of the Bug Tracker's features and i have used OOP and some ES6 classes to generate user objects and bug objects. I have mostly focused in this first instance on the Login Form which i have begun to add functionality for.