Communicating with database and core functionality
over the last couple days i have been further implementing the functionality that requires the database to operate such as updating users bugs array or projects list on the firebase firestore.  i have also implemented a way of giving this data back the classes it was initialized with as firestore seems to have trouble storing custom objects. To get around this i send the data to firebase in JSON and convert it back to a new custom object when the data is recieved so that i can access the custom properties and methods of the object prototype. Once the firestore was able to update existing user data i got to work implementing the add bug functionality and had it be linked to the project name so that the user can have bugs for seperate projects and have them be rendered to the screen upon selecting that project. i also implemented the bugs being rendered to the projects bugs list.