Building for multiperson use
In this build i have fleshed out almost all functionality for the site and implemented a means for other users to share and see each others bugs with their teamcode. I have the user enter a code to either create or join. should they create one their teamcode will be stored to firebase as their input. If they join the state.accounts array is searched for any other users with the same teamcode and then pulls all their unique projects and bugs. I have chosen not to save this data as it may result in some storage concerns so upon login the user will need to input the teamcode again. This however will allow the user to use multiple teamcodes whenever they choose. Some styling upgrades have been added as well to some buttons and other UI improvements such as adding a button to take the user out of the specific project view.