Opening and Editing Bugs
In this update i have implemented the ability for a user to click on a bug and have a details panel appear on screen where the user can see all information about the bug, i did run into some trouble when trying to figure out which bug was clicked on and ive had to use a rather ugly way of using methods on the text content to extrapolate the name of the bug and find it from that. I have also added the ability to edit existing bugs and update any individual piece of the data and have the existing bug replaced in storage with the edited version. to do this i actually refactored the controlAddBug function as a lot of the functionality for editing the bug is already there in the add bug function. so i abstracted some of the code into a new function that i am now re using to retrieve, validate and then create a new Bug object