Final features and functionality
I have been working on implementing the final features and functionality. The main things are the implementation of a search bar which will allow the user to search their bugs using either the name or the tags. The user does not need the entire thing just an accurate portion of the name or tag for example if the name was New Bug then simply Bug in the search bar would return this bug. I have implemented full responsiveness using media queries to work with a viewport as small as the iphone SE(the smalles example given through chrome devtools). I have also implemented the ability for users to delete an entire project along with the bugs it contains, the user is prompted to confirm their decision before it executes. Other than this i have fixed some bugs and cleaned some other code up. There is still alot of refactoring to be done and small details such as fav-icon etc and some test data for people to see and use that need to be implemented and then the application will be complete.