Bug Fixing and final styling
In this update i did some extensive testing on features to try to find any bugs or issues with the functionality. The only major issue found was a bug where if a user created an account and in that same session joined a team and made a bug for it the bug would not be openable until the user had refreshed the page. This has been patched. Everything else was minor and easily fixed and the styles were all updated to their final states. This is the final product at this stage. I have some planned features that i may implement such as a dashboard or a trello board style view for teams of users. There will be a final refactoring update in the near future.